requirements.txt
asgiref==3.8.1
django==5.2.1
djangorestframework==3.16.0
pip==25.0.1
sqlparse==0.5.3
python-dotenv==1.0.0